Reorganize protvista data repository
