+++ app/assets/javascripts/3dbio_viewer/src/data/repositories/protvista/tracks/variants.ts
@@ -1,4 +1,4 @@
-import { Variant, VariantFilter, Variants } from "../../../domain/entities/Variant";
+import { Variant, VariantFilter, Variants } from "../../../../domain/entities/Variant";
 
 export interface EbiVariation {
     accession: string; // "P0DTC2";
